Enable to filter events by any sortoption and properly type sort argument
