+++ changelog.md
@@ -91,7 +91,7 @@ Our Roadmap ahead:
 
 ## Major
 
-- [api] `query.events` is now extended with additional filter parameters `create` get events created after the provided event and the previous `type` field is changed to `types` and accepts array of event types. in addition it is now possible to sort the results based on `created` time or event `type`
+- [api] `query.events` is now extended with additional filter parameters `create` get events created after the provided event and the previous `type` field is changed to `types` and accepts array of event types. in addition it is now possible to `sort` the results any field on event type
 - [core] Bulk Importer now supports skipping the built-in cache invalidation run at the end by using
   `skipCacheInvalidation`
 - [core] Bulk Importer now supports defining custom entities to sync
