+++ packages/api/src/schema/types/product/bundle-product.js
@@ -17,7 +17,7 @@ export default [
       reviews(
         limit: Int = 10
         offset: Int = 0
-        sort: [ProductReviewSortOptionInput!]
+        sort: [SortOptionInput!]
         queryString: String
       ): [ProductReview!]!
