+++ packages/api/src/schema/types/product/configurable-product.js
@@ -82,7 +82,7 @@ export default [
       reviews(
         limit: Int = 10
         offset: Int = 0
-        sort: [ProductReviewSortOptionInput!]
+        sort: [SortOptionInput!]
         queryString: String
       ): [ProductReview!]!
       catalogPriceRange(
