+++ packages/api/src/schema/types/product/plan-product.js
@@ -59,7 +59,7 @@ export default [
       reviews(
         limit: Int = 10
         offset: Int = 0
-        sort: [ProductReviewSortOptionInput!]
+        sort: [SortOptionInput!]
         queryString: String
       ): [ProductReview!]!
       plan: ProductPlanConfiguration
