+++ packages/api/src/schema/types/product/simple-product.js
@@ -42,7 +42,7 @@ export default [
       reviews(
         limit: Int = 10
         offset: Int = 0
-        sort: [ProductReviewSortOptionInput!]
+        sort: [SortOptionInput!]
         queryString: String
       ): [ProductReview!]!
     }
